requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
binary2strings==0.1.13
blinker==1.9.0
cachetools==6.2.6
captcha==0.7.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorclass==2.2.2
contourpy==1.3.3
cryptography==46.0.5
cxxfilt==0.3.0
cycler==0.12.1
deprecation==2.1.0
distro==1.9.0
dncil==1.0.2
filetype==1.2.0
flare-floss==3.1.1
Flask==3.1.3
Flask-HTTPAuth==4.8.0
Flask-WTF==1.2.2
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
funcy==2.0
google-auth==2.48.0
google-genai==1.64.0
gunicorn==25.1.0
h11==0.16.0
h2==4.3.0
halo==0.0.31
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
intervaltree==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.2.10
langchain-core==1.2.14
langchain-google-genai==4.2.1
langsmith==0.7.6
LnkParse3==1.6.0
log-symbols==0.0.14
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
msgpack==1.1.2
msoffcrypto-tool==6.0.0
multidict==6.7.1
networkx==3.1
numpy==2.4.2
olefile==0.47
oletools==0.60.2
orjson==3.11.7
packaging==26.0
pefile==2024.8.26
pillow==12.1.1
pipdeptree==2.31.0
platformdirs==4.9.2
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyelftools==0.32
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-magic==0.4.27
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
semantic-version==2.10.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
spinners==0.0.24
StrEnum==0.4.15
tabulate==0.9.0
tenacity==9.1.4
termcolor==3.3.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid_utils==0.14.1
viv_utils==0.8.1
vivisect==1.3.0
wcwidth==0.6.0
websockets==15.0.1
Werkzeug==3.1.6
WTForms==3.2.1
xmltodict==1.0.4
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
